+++ packages/terra-date-time-picker/tests/wdio/date-time-picker-spec.js
@@ -211,6 +211,7 @@ Terra.describeViewports('DateTimePicker', ['tiny', 'large'], () => {
       expect($('#ambiguous-date').getText()).toEqual('No');
       expect($('#complete-date').getText()).toEqual('Yes');
       expect($('#valid-date').getText()).toEqual('Yes');
+      expect($('#date-time').getText()).toEqual('{"date":"01","month":"05","year":"2019","hour":"10","minute":"10"}');
     });
   });
 
@@ -832,12 +833,16 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           $('input[name="terra-time-hour-input"]').click();
           browser.keys('-');
 
-          const today = moment.utc().subtract(1, 'minutes');
+          const today = moment.utc().subtract(1, 'hours');
           expect($('input[name="terra-date-year-input"]')).toHaveValue(today.year().toString());
           expect($('input[name="terra-date-month-input"]')).toHaveValue((`0${(today.month() + 1)}`).slice(-2));
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          const hourValue = Number($('input[name="terra-time-hour-input"]').getValue());
+          const hourTimeValue = today.hours();
+          const hourInRange = (hourTimeValue === hourValue + 1 || hourTimeValue === hourValue);
+          expect(hourInRange).toBe(true);
         });
 
         it('should set date to today and now minus 1 minute in the minute input', () => {
@@ -852,6 +857,10 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
+          const minuteTimeValue = today.minutes();
+          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          expect(minuteInRange).toBe(true);
         });
 
         it('should set date to today and now minus 1 minute in the second input', () => {
@@ -866,10 +875,6 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
-          const secondsValue = Number($('input[name="terra-time-second-input"]').getValue());
-          const secondsTimeValue = today.seconds();
-          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
-          expect(secondsInRange).toBe(true);
         });
       });
 
@@ -908,6 +913,10 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
+          const minuteTimeValue = today.minutes();
+          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          expect(minuteInRange).toBe(true);
         });
 
         it('should set date to today and now minus 1 minute in the second input', () => {
@@ -926,10 +935,6 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
-          const secondsValue = Number($('input[name="terra-time-second-input"]').getValue());
-          const secondsTimeValue = today.seconds();
-          const secondsInRange = (secondsTimeValue === secondsValue + 1 || secondsTimeValue === secondsValue);
-          expect(secondsInRange).toBe(true);
         });
       });
 
@@ -1151,12 +1156,16 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           $('input[name="terra-time-hour-input"]').click();
           browser.keys('+');
 
-          const today = moment.utc().add(1, 'minutes');
+          const today = moment.utc().add(1, 'hours');
           expect($('input[name="terra-date-year-input"]')).toHaveValue(today.year().toString());
           expect($('input[name="terra-date-month-input"]')).toHaveValue((`0${(today.month() + 1)}`).slice(-2));
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          const hourValue = Number($('input[name="terra-time-hour-input"]').getValue());
+          const hourTimeValue = today.hours();
+          const hourInRange = (hourTimeValue === hourValue + 1 || hourTimeValue === hourValue);
+          expect(hourInRange).toBe(true);
         });
 
         it('should set date to today and now plus 1 minute in the minute input', () => {
@@ -1171,6 +1180,10 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
+          const minuteTimeValue = today.minutes();
+          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          expect(minuteInRange).toBe(true);
         });
 
         it('should set date to today and now plus 1 minute in the second input', () => {
@@ -1227,6 +1240,10 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
           expect($('input[name="terra-date-day-input"]')).toHaveValue((`0${today.date()}`).slice(-2));
           expect($('input[name="terra-time-hour-input"]')).toHaveValue((`0${today.hour()}`).slice(-2));
           expect($('input[name="terra-time-minute-input"]')).toHaveValue((`0${today.minute()}`).slice(-2));
+          const minuteValue = Number($('input[name="terra-time-minute-input"]').getValue());
+          const minuteTimeValue = today.minutes();
+          const minuteInRange = (minuteTimeValue === minuteValue + 1 || minuteTimeValue === minuteValue);
+          expect(minuteInRange).toBe(true);
         });
 
         it('should set date to today and now plus 1 minute in the second input', () => {
@@ -1993,3 +2010,4 @@ Terra.describeViewports('DateTimePicker', ['large'], () => {
     });
   });
 });
+