+++ packages/terra-framework-docs/src/terra-dev-site/doc/time-input/example/TimeInputNoDefaultTwelveHour.jsx
@@ -2,23 +2,30 @@ import React from 'react';
 import TimeInput from 'terra-time-input/lib/TimeInput';
 import TimeUtil from 'terra-time-input/lib/TimeUtil';
 import Field from 'terra-form-field';
+import moment from 'moment-timezone';
 
 class timeInput extends React.Component {
   constructor(props) {
     super(props);
-    this.state = { time: '' };
+    this.state = { time: '', labelValue: '' };
     this.handleTimeChange = this.handleTimeChange.bind(this);
   }
 
-  handleTimeChange(event, time) {
-    this.setState({ time });
+  handleTimeChange(event, time, meridiem) {
+    if (meridiem === 'p.m.') {
+      const updatedTime = moment(time, ['HH:mm']).format('hh:mm');
+      this.setState({ labelValue: updatedTime });
+    } else {
+      this.setState({ labelValue: time });
+      this.setState({ time });
+    }
   }
 
   render() {
     return (
       <div>
         <Field
-          label={`Enter Time: ${this.state.time}`}
+          label={`Enter Time: ${this.state.labelValue}`}
           htmlFor="time-input-value"
         >
           <TimeInput